+++ packages/test-utils/src/index.ts
@@ -58,6 +58,8 @@ export const expecter = (
 
       // for better error message snapshots
       noErrorTruncation: true,
+
+      types: ['node'],
     })
 
     const snippets = tsSnippet.snippet(