+++ packages/theme-ui/test/color-modes.tsx
@@ -38,6 +38,7 @@ test('renders with color modes', () => {
   let rendered:
     | renderer.ReactTestRendererJSON
     | renderer.ReactTestRendererJSON[]
+    | null = null
 
   const Mode = () => {
     const [colorMode] = useColorMode()
@@ -508,6 +509,7 @@ test('raw color values are passed to theme-ui context when custom properties are
 })
 
 test('warns when localStorage is disabled', () => {
+  const restore = mockConsole()
   window.matchMedia = jest.fn().mockImplementation((query) => {
     return {
       matches: false,
@@ -534,4 +536,6 @@ test('warns when localStorage is disabled', () => {
   )
 
   expect(mode).toBe(undefined)
+  expect(console.warn).toHaveBeenCalled()
+  restore()
 })